pc-controller/pyproject.toml
[project]
name = "pc-controller"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.121.2",
    "loguru>=0.7.3",
    "psutil>=7.1.3",
    "python-dotenv>=1.2.1",
]
